+++ src/nmath/bessel_j.c
@@ -61,7 +61,7 @@ double bessel_j(double x, double alpha)
                 ((alpha == na) ? 0 : bessel_y(x, -alpha) * sin(M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)(nb - 1);
 #ifdef MATHLIB_STANDALONE
     bj = (double *)calloc(nb, sizeof(double));
     if (!bj)
@@ -77,7 +77,7 @@ double bessel_j(double x, double alpha)
             MATHLIB_WARNING4(_("bessel_j(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else
-            MATHLIB_WARNING2(_("bessel_j(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_j(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = bj[nb - 1];
 #ifdef MATHLIB_STANDALONE
@@ -114,7 +114,7 @@ double bessel_j_ex(double x, double alpha, double *bj)
                 ((alpha == na) ? 0 : bessel_y_ex(x, -alpha, bj) * sin(M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)(nb - 1);
     J_bessel(&x, &alpha, &nb, bj, &ncalc);
     if (ncalc != nb)
     { /* error input */
@@ -122,7 +122,7 @@ double bessel_j_ex(double x, double alpha, double *bj)
             MATHLIB_WARNING4(_("bessel_j(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else
-            MATHLIB_WARNING2(_("bessel_j(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_j(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = bj[nb - 1];
     return x;
