+++ src/nmath/bessel_y.c
@@ -60,7 +60,7 @@ double bessel_y(double x, double alpha)
                 ((alpha == na) ? 0 : bessel_j(x, -alpha) * sin(M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)nb-1);
 #ifdef MATHLIB_STANDALONE
     by = (double *)calloc(nb, sizeof(double));
     if (!by)
@@ -78,7 +78,7 @@ double bessel_y(double x, double alpha)
             MATHLIB_WARNING4(_("bessel_y(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else /* ncalc >= 0 */
-            MATHLIB_WARNING2(_("bessel_y(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_y(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = by[nb - 1];
 #ifdef MATHLIB_STANDALONE
@@ -115,7 +115,7 @@ double bessel_y_ex(double x, double alpha, double *by)
                 ((alpha == na) ? 0 : bessel_j_ex(x, -alpha, by) * sin(M_PI * alpha)));
     }
     nb = 1 + (long)na; /* nb-1 <= alpha < nb */
-    alpha -= (nb - 1);
+    alpha -= (double)(nb - 1);
     Y_bessel(&x, &alpha, &nb, by, &ncalc);
     if (ncalc != nb)
     { /* error input */
@@ -125,7 +125,7 @@ double bessel_y_ex(double x, double alpha, double *by)
             MATHLIB_WARNING4(_("bessel_y(%g): ncalc (=%ld) != nb (=%ld); alpha=%g. Arg. out of range?\n"), x, ncalc, nb,
                              alpha);
         else /* ncalc >= 0 */
-            MATHLIB_WARNING2(_("bessel_y(%g,nu=%g): precision lost in result\n"), x, alpha + nb - 1);
+            MATHLIB_WARNING2(_("bessel_y(%g,nu=%g): precision lost in result\n"), x, alpha + (double)nb - 1);
     }
     x = by[nb - 1];
     return x;
